requirements.txt
flask==3.1.2
reportlab==4.4.3
requests==2.32.5
python-dotenv==1.1.1